# coding-openapi.yaml
openapi: "3.1.0"
info:
  title: "GPT-API Coding Agent API"
  version: "1.0.0"
  description: "Narrow Custom GPT Coding Agent API with strict state-machine contracts."
paths:
...
  /github/issue/read:
    post:
      summary: "Issue Read"
      operationId: "issue_read_github_issue_read_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubIssueReadRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/read:
    post:
      summary: "Pr Read"
      operationId: "pr_read_github_pr_read_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPrReadRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/checks/read:
    post:
      summary: "Checks Read"
      operationId: "checks_read_github_checks_read_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubChecksReadRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/comment:
    post:
      summary: "Pr Comment"
      operationId: "pr_comment_github_pr_comment_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPrCommentRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/create-from-task:
    post:
      summary: "Pr Create From Task"
      operationId: "pr_create_from_task_github_pr_create_from_task_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPrCreateFromTaskRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/checks/diagnose:
    post:
      summary: "Checks Diagnose"
      operationId: "checks_diagnose_github_checks_diagnose_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubChecksDiagnoseRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/apply-feedback-plan:
    post:
      summary: "Apply Feedback Plan"
      operationId: "apply_feedback_plan_github_pr_apply_feedback_plan_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubApplyFeedbackPlanRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/update-body:
    post:
      summary: "Pr Update Body"
      operationId: "pr_update_body_github_pr_update_body_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubPrUpdateBodyRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/review-comments:
    post:
      summary: "Pr Review Comments"
      operationId: "pr_review_comments_github_pr_review_comments_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubReviewCommentsRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/checks/logs:
    post:
      summary: "Checks Logs"
      operationId: "checks_logs_github_checks_logs_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubChecksLogsRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/branch/push:
    post:
      summary: "Branch Push"
      operationId: "branch_push_github_branch_push_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubBranchPushRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/checks/repair-plan:
    post:
      summary: "Checks Repair Plan"
      operationId: "checks_repair_plan_github_checks_repair_plan_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubCiRepairPlanRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
  /github/pr/feedback-to-patch-contract:
    post:
      summary: "Feedback To Patch Contract"
      operationId: "feedback_to_patch_contract_github_pr_feedback_to_patch_contract_post"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GithubFeedbackContractRequest"
        required: true
      responses:
        200:
          description: "Successful Response"
          content:
            application/json:
              schema:

        422:
          description: "Validation Error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
      security:
        -
          ApiKeyAuth:
...
components:
  schemas:
...
    GithubApplyFeedbackPlanRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        comments:
          items:
            additionalProperties: true
            type: "object"
          type: "array"
          title: "Comments"
      type: "object"
      required:
        - "workspace_path"
        - "comments"
      title: "GithubApplyFeedbackPlanRequest"
    GithubBranchPushRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        remote:
          type: "string"
          title: "Remote"
          default: "origin"
        branch:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Branch"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
      title: "GithubBranchPushRequest"
    GithubChecksDiagnoseRequest:
      properties:
        checks:
          items:
            additionalProperties: true
            type: "object"
          type: "array"
          title: "Checks"
      type: "object"
      required:
        - "checks"
      title: "GithubChecksDiagnoseRequest"
    GithubChecksLogsRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        ref:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Ref"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
      title: "GithubChecksLogsRequest"
    GithubChecksReadRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        ref:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Ref"
      type: "object"
      required:
        - "workspace_path"
      title: "GithubChecksReadRequest"
    GithubCiRepairPlanRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        checks:
          anyOf:
            -
              items:
                additionalProperties: true
                type: "object"
              type: "array"
            -
              type: "null"
          title: "Checks"
        logs:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Logs"
      type: "object"
      required:
        - "workspace_path"
      title: "GithubCiRepairPlanRequest"
    GithubFeedbackContractRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        comments:
          items:
            additionalProperties: true
            type: "object"
          type: "array"
          title: "Comments"
      type: "object"
      required:
        - "workspace_path"
        - "comments"
      title: "GithubFeedbackContractRequest"
    GithubIssueReadRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        issue:
          type: "string"
          title: "Issue"
      type: "object"
      required:
        - "workspace_path"
        - "issue"
      title: "GithubIssueReadRequest"
    GithubPrCommentRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
        body:
          type: "string"
          title: "Body"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
        - "pr"
        - "body"
      title: "GithubPrCommentRequest"
    GithubPrCreateFromTaskRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        task_id:
          type: "string"
          title: "Task Id"
        title:
          anyOf:
            -
              type: "string"
            -
              type: "null"
          title: "Title"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
        - "task_id"
      title: "GithubPrCreateFromTaskRequest"
    GithubPrReadRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
      type: "object"
      required:
        - "workspace_path"
        - "pr"
      title: "GithubPrReadRequest"
    GithubPrUpdateBodyRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
        body:
          type: "string"
          title: "Body"
        dry_run:
          type: "boolean"
          title: "Dry Run"
          default: true
      type: "object"
      required:
        - "workspace_path"
        - "pr"
        - "body"
      title: "GithubPrUpdateBodyRequest"
    GithubReviewCommentsRequest:
      properties:
        workspace_path:
          type: "string"
          title: "Workspace Path"
        pr:
          type: "string"
          title: "Pr"
      type: "object"
      required:
        - "workspace_path"
        - "pr"
      title: "GithubReviewCommentsRequest"
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              -
                type: "string"
              -
                type: "integer"
          type: "array"
          title: "Location"
        msg:
          type: "string"
          title: "Message"
        type:
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
...
  securitySchemes:
    ApiKeyAuth:
      type: "apiKey"
      in: "header"
      name: "x-api-key"
servers:
  -
    url: "https://gpt-api.ngrok.app"
security:
  -
    ApiKeyAuth: